Reinforce and Protect ajax

+++ include/ajax/ajax_remove_submenu.php
@@ -26,5 +26,6 @@ if ( ! defined ('ALLOWED') ) die('Appel direct ne sont pas permis');
 if ($g_user->check_module('CFGPRO')==0)
     die();
 $p_profile_menu_id=HtmlInput::default_value_get('p_profile_menu_id', 0);
+if ( $p_profile_menu_id == 0 ||isNumber($p_profile_menu_id)==0)    throw new Exception(_('Donnée invalide'));
 $cn->exec_sql('delete from profile_menu where pm_id = $1 or pm_id_dep=$1',array($p_profile_menu_id))
 ?>
